feat: Phases 6-17 (Meridian, Prism, Chronos, Plexus, Canopy, Flux, Synapse, Mirror, Harbor connectors, Canvas AgentMonitor) and refresh README/CLAUDE.md
Adds the six SQL-extension engines and the Synapse/Mirror agent layers to
tpt-keystone, fills in the remaining Harbor source connectors and web
dashboard, adds Canvas's AgentMonitor component, and rewrites README.md and
CLAUDE.md to describe the full multi-engine platform instead of only
Keystone.

+++ tpt-harbor/src/sources/mongodb.rs
@@ -294,7 +294,7 @@ impl MongoConn {
         let req_id = self.request_id;
         self.request_id += 1;
 
-        self.write_buf.put_i32_le(msg_len); // message length
+        self.write_buf.put_i32_le(msg_len as i32); // message length
         self.write_buf.put_i32_le(req_id);  // request id
         self.write_buf.put_i32_le(0);       // response to (0 = request)
         self.write_buf.put_i32_le(2013);    // opCode: OP_MSG = 2013
